+++ packages/assets/icons/index.js
@@ -24,6 +24,8 @@ export default {
     maintenance: 'mdi-wrench-outline',
     shopfloorDashboard: 'mdi-view-dashboard-outline',
     utilityDashboard: 'mdi-flash-outline',
+    energyDashboard: 'mdi-flash-outline',
+    utility: 'mdi-flash-outline',
     processDashboard: 'mdi-gauge',
     poProductionDashboard: 'mdi-view-dashboard-outline',
     productionLog: 'mdi-circle-edit-outline',