+++ packages/infinity-modules/product-management/src/components/dialogs/AddProduct.vue
@@ -51,6 +51,7 @@
             :label="$t('displayTags.productTypeDescription')"
             prepend-icon="mdi-tray-plus"
             v-model="product.description"
+            required
           ></v-text-field>
           <v-text-field
             :disabled="saving"
@@ -104,6 +105,7 @@
             :disabled="!lineSelected ||
             !selectedRoadmap ||
             !product ||
+            !selectedBom ||
             !valid"
             @click="saveProduct"
             >
@@ -171,7 +173,7 @@ export default {
   methods: {
     ...mapMutations('helper', ['setAlert']),
     ...mapMutations('productManagement', ['setAddProductDialog']),
-    ...mapActions('productManagement', ['createProduct']),
+    ...mapActions('productManagement', ['createProduct', 'checkDuplicateProduct']),
     async saveProduct() {
       this.$refs.form.validate();
       if (!this.product.productname) {
@@ -180,11 +182,16 @@ export default {
           type: 'error',
           message: 'PRODUCT_NAME_EMPTY',
         });
+      } else if (!this.product.description) {
+        this.setAlert({
+          show: true,
+          type: 'error',
+          message: 'PRODUCT_DESCRIPTION_EMPTY',
+        });
       } else {
-        const duplicateProduct = this.productList.filter(
-          (o) => o.productname.toLowerCase().split(' ').join('') === this.product.productname.toLowerCase().split(' ').join(''),
-        );
-        if (duplicateProduct.length > 0) {
+        const query = `?query=productname=="${this.product.productname}"`;
+        const duplicateProduct = await this.checkDuplicateProduct(query);
+        if (duplicateProduct) {
           this.product.productname = '';
           this.setAlert({
             show: true,