+++ packages/infinity-modules/report-viewer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.3](https://bitbucket.org/entrib/shopworx/compare/v4.3.2...v4.3.3) (2021-05-26)
+
+
+### Bug Fixes
+
+* **report-viewer:** fixed chart display on report change ([69f8ee6](https://bitbucket.org/entrib/shopworx/commits/69f8ee63cbf8ab7786dbc746991273868b042bbc))
+
+
+### Features
+
+* **locale:** added utility reports locale ([988d268](https://bitbucket.org/entrib/shopworx/commits/988d268142c77c1d92152335ffaa888517db4981))
+
+
+
+
+
 ## [4.3.2](https://bitbucket.org/entrib/shopworx/compare/v4.3.1...v4.3.2) (2021-05-18)
 
 **Note:** Version bump only for package @shopworx/report-viewer